requirements.txt
django~=3.1.3
djangorestframework~=3.12.2
graphene~=2.1.8
pandas~=1.1.4
channels~=3.0.2
bokeh~=2.2.3
celery~=5.0.2
numpy~=1.19.4
prefect~=0.13.16
django-crispy-forms~=1.10.0
plotly~=4.12.0
dash~=1.17.0
panel~=0.10.2
param~=1.10.0
ipyaggrid~=0.2.1
django-webpack-loader==0.6.0